@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "thothglyph-doc"
-version = "0.1.3"
+version = "0.2.0"
 description = "A Documentation converter and language for Engineers"
 authors = ["nakandev <nakandev.s@gmail.com>"]
 license = "MIT"
@@ -13,16 +13,17 @@ packages = [
 thothglyph = "thothglyph.app.converter:main"
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8"
 pillow = "^9.2.0"
 cairosvg = "^2.5.2"
+markdown-it-py = {extras = ["plugins"], version = "^3.0.0"}
 
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^0.982"
 
 [tool.mypy]
-python_version = "3.7"
+python_version = "3.8"
 ignore_missing_imports = true
 check_untyped_defs = true
 
