@@ -7,10 +7,14 @@ from logging import Logger, StreamHandler, Formatter  # type: ignore
 NAMESPACE = 'thothglyph'
 
 
-def getLogger(modname) -> Logger:
-    logger = logging.getLogger(NAMESPACE + '.' + modname)
-    handler = StreamHandler()
-    formatter = Formatter('[%(levelname)s] %(message)s')
-    handler.setFormatter(formatter)
-    logger.addHandler(handler)
+def getLogger(modname=None) -> Logger:
+    name = NAMESPACE
+    if modname:
+        name += '.' + modname
+    logger = logging.getLogger(name)
+    if not modname:
+        handler = StreamHandler()
+        formatter = Formatter('[%(levelname)s] %(message)s')
+        handler.setFormatter(formatter)
+        logger.addHandler(handler)
     return logger
