@@ -3,12 +3,13 @@ import importlib
 
 readerclass = {
     'tglyph': 'TglyphReader',
+    'md': 'MdReader',
 }
 
 
-def Reader(target: str):
+def ReaderClass(target: str):
     modname = target
     clsname = readerclass[target]
     modpath = 'thothglyph.reader.{}'.format(modname)
     module = importlib.import_module(modpath)
-    return getattr(module, clsname)()
+    return getattr(module, clsname)
